pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asr-eval-system"
version = "0.1.0"
description = "Undergraduate-friendly speech recognition performance measurement system"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Codex" }]
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
app = [
  "streamlit>=1.40",
  "pyside6>=6.7",
  "psutil>=5.9",
]
real_models = [
  "faster-whisper>=1.1.0",
  "paddlespeech>=1.4.0",
  "psutil>=5.9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

